// src/app/costing/model/masterpriceentry.ts
export class Price {
  itempriceid: any;
  rebate: any;
  commission: any;
  extrarebate: any;
  price: any;
  length: any;
  width: any;
  stockratedate: any;
  userid: any;
  buyer: any;
  subbuyer: any;
  sizewiseqty: any;
  sizewisevalue: any;
  rebatepercent: boolean;
  commissionpercent: boolean;
  extrarebatepercent: boolean;
  salespersonid: any;
  customerid: any;
  sampleid: any;
  partnoid: any;
  cost: any;
  finishtype: any;
  minordervalue: any;
  createddate: any;
  minorderqty: any;
  productcategoryid: any;
  perquantity: any;
  statusid: any;
  cvrate: any;
  createdby: any;
  merchandisercommission: any;
  merchandisercommpercent: boolean;
  hmpercentage: boolean;
  remarks: any;
  constructor() {
    this.stockratedate = new Date();
    this.createddate = new Date();
    this.itempriceid = 0;
    this.rebate = 0;
    this.commission = 0;
    this.extrarebate = 0;
    this.price = 0;
    this.length = 0;
    this.width = 0;
    this.userid = 0;
    this.buyer = 0;
    this.subbuyer = 0;
    this.salespersonid = 0;
    this.customerid = 0;
    this.sampleid = 0;
    this.partnoid = 0;
    this.createdby = sessionStorage.getItem('empid');
    this.cost = 0;
    this.finishtype = -1;
    this.minordervalue = 0;
    this.minorderqty = 0;
    this.productcategoryid = 0;
    this.perquantity = 1000;
    this.sizewiseqty = 0;
    this.sizewisevalue = 0;
    this.merchandisercommission = 0;
    this.statusid = 0;
    this.cvrate = 0;
    this.rebatepercent = false;
    this.commissionpercent = false;
    this.extrarebatepercent = false;
    this.merchandisercommpercent = false;
  }
}